Move errors to their own file
